@@ -12,13 +12,14 @@
     "homepage": "https://github.com/roromix/SpreadsheetBundle",
     "license": "MIT",
     "require": {
-        "symfony/framework-bundle": "~3.2|~4.0|~5.0|~6.0|~7.0",
+        "php": ">=7.1",
+        "symfony/framework-bundle": "~4.0|~5.0|~6.0|~7.0",
         "phpoffice/phpspreadsheet": "~1.0"
     },
     "require-dev": {
-        "symfony/symfony": "~3.2|~4.0|~5.0|~6.0|~7.0",
-        "symfony/phpunit-bridge": "~3.2|~4.0|~5.0|~6.0|~7.0",
-        "sensio/framework-extra-bundle": "~3.0|~4.0|~5.0|~6.0"
+        "symfony/symfony": "~4.0|~5.0|~6.0|~7.0",
+        "symfony/phpunit-bridge": "^5.3|~6.0|~7.0",
+        "sensio/framework-extra-bundle": "~5.0|~6.0"
     },
     "autoload": {
         "psr-4": { "Roromix\\Bundle\\SpreadsheetBundle\\": "src/" }
