@@ -10,17 +10,17 @@ class Configuration implements ConfigurationInterface
     /**
      * {@inheritDoc}
      */
-    public function getConfigTreeBuilder()
+    public function getConfigTreeBuilder(): TreeBuilder
     {
         $treeBuilder = new TreeBuilder('roromix_spreadsheet');
+
         if (method_exists($treeBuilder, 'getRootNode')) {
             $rootNode = $treeBuilder->getRootNode();
         } else {
             // BC layer for symfony/config 4.1 and older
             $rootNode = $treeBuilder->root('roromix_spreadsheet');
         }
-        
-        
+
         return $treeBuilder;
     }
 }
